requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-PyMongo==2.3.0
heyoo==0.1.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pymongo==4.6.3
python-dotenv==1.0.1
pytz==2024.1
pywa==1.16.2
requests==2.31.0
requests-toolbelt==1.0.0
typing==3.7.4.3
urllib3==2.2.1
Werkzeug==3.0.1
